re-update

+++ api/v1/views/amenities.py
@@ -18,7 +18,6 @@ def all_my_amenities():
         ls = []
         for key in am:
             ls.append(key.to_dict())
-
     return jsonify(ls)
 
 
@@ -26,12 +25,12 @@ def all_my_amenities():
                  methods=['GET'], strict_slashes=False)
 def get_amenity(amenity_id):
     """ retrive amenity object """
-
-    am = storage.get(Amenity, amenity_id)
-    if am is None:
-        abort(404)
-
-    return jsonify(am.to_dict())
+    if request.method == 'GET':
+        am = storage.get(Amenity, amenity_id)
+        if am is None:
+            abort(404)
+        else:
+            return jsonify(am.to_dict())
 
 
 @app_views.route('/amenities/<amenity_id>',
@@ -39,30 +38,31 @@ def get_amenity(amenity_id):
 def delete_amenity_by_id(amenity_id):
     """ delete a amenity object """
 
-    am = storage.get(Amenity, amenity_id)
-
-    if am is None:
-        abort(404)
-
-    storage.delete(am)
-    storage.save()
-    return jsonify({})
+    if request.method == 'DELETE':
+        am = storage.get(Amenity, amenity_id)
+        if am is None:
+            abort(404)
+        else:
+            storage.delete(am)
+            storage.save()
+            return jsonify({}), 200
 
 
 @app_views.route('/amenities', methods=['POST'], strict_slashes=False)
 def create_amenity():
     """ create Amenity """
+    if request.method == 'POST':
+        req_name = request.get_json()
+        req = request.headers.get('Content-Type')
+        if req != 'application/json':
+            return jsonify('Not a JSON'), 400
 
-    req = request.get_json()
-    if req is None:
-        abort(400, 'Not a JSON')
+        if req_name('name') is None:
+            return jsonify ('Missing name'), 400
 
-    if req.get('name') is None:
-        abort(400, 'Missing name')
-
-    am = Amenity(**req)
-    am.save()
-    return jsonify(am.to_dict()), 201
+        am = Amenity(**req)
+        am.save()
+        return jsonify(am.to_dict()), 201
 
 
 @app_views.route('/amenities/<amenity_id>',
@@ -72,17 +72,13 @@ def update_amenity(amenity_id):
 
     if request.method == 'PUT':
         am = storage.get(Amenity, amenity_id)
-        if amenity is None:
-            abort(404)
-
-        type_req = request.headers.get('Content-Type')
-        if type_req != 'application/json':
-            abort(400, "Not a JSON")
-
-        req = request.get_json()
+        req = request.headers.get('Content-Type')
+        if req != 'application/json':
+            return jsonify('Not a JSON'), 400
+        req_name = request.get_json()
         if am is not None:
-            if 'name' in req:
-                am.name = req['name']
+            if 'name' in req_name:
+                am.name = req_name['name']
                 am.save()
-                return (jsonify(am.to_dict()), 200)
+                return jsonify(am.to_dict()), 200
         abort(404)
